test(dynamicentity): get, update, create, delete
